+++ src/main/java/org/magicdgs/readtools/tools/barcodes/BamBarcodeDetector.java
@@ -33,8 +33,10 @@ import org.magicdgs.readtools.tools.ReadToolsBaseTool;
 import org.magicdgs.readtools.tools.barcodes.dictionary.BarcodeDictionary;
 import org.magicdgs.readtools.tools.barcodes.dictionary.BarcodeDictionaryFactory;
 import org.magicdgs.readtools.tools.barcodes.dictionary.decoder.BarcodeDecoder;
+import org.magicdgs.readtools.tools.barcodes.dictionary.decoder.stats.MatcherStat;
 import org.magicdgs.readtools.utils.fastq.RTFastqContstants;
 import org.magicdgs.readtools.utils.logging.ProgressLoggerExtension;
+import org.magicdgs.readtools.utils.misc.Formats;
 import org.magicdgs.readtools.utils.misc.IOUtils;
 import org.magicdgs.readtools.utils.record.SAMRecordUtils;
 
@@ -50,6 +52,7 @@ import org.broadinstitute.hellbender.exceptions.UserException;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Collection;
 import java.util.HashSet;
 
 /**
@@ -125,9 +128,12 @@ public final class BamBarcodeDetector extends ReadToolsBaseTool {
             progress.record(record);
         });
         progress.logNumberOfVariantsProcessed();
-        decoder.logMatcherResult(logger);
+        // output statistics and log them
         try {
-            decoder.outputStats(IOUtils.makeMetricsFile(outputPrefix).toFile());
+            final Collection<MatcherStat> stats = decoder
+                    .outputStats(IOUtils.makeMetricsFile(outputPrefix).toFile());
+            stats.forEach(s -> logger.info("Found {} records for {} ({}).",
+                    Formats.commaFmt.format(s.RECORDS), s.SAMPLE, s.BARCODE));
         } catch (IOException e) {
             throw new UserException(e.getMessage(), e);
         }